Fix installer permission issues
As of c5608b59b14afc0cefd724a2181f15f15584c8bc mailcatcher is installed
as a global gem, rather than in the application Gemfile bundle.

`script/install-as-user` is run as the application user, so doesn't have
permission to install gems globally.

This commit moves the install of mailcatcher to
`script/site-specific-install.sh`, which is run as root.

Also only installs the gem when `$DEVELOPMENT_INSTALL` is set.