Refactor classification controller specs
Update the post_status helper method to:
1. Use the info_request defined in the let method by default instead of
   needing this passed in as an argument
2. Take an optional message argument
3. Replaces the expect_redirect helper so the redirect to expectation is
   more explicit in each spec
3. Be used throughout the file. Only isn't used when certain params have
   to be omitted.

Additionally this uses factories instead of mocking Time to make a info
request overdue or very overdue.

I've added missing specs for when to show or not to show a flash notice
where applicable.

There are a few session assignments removed as these have been already
assigned in other places in the file.

Overall these changes make the file more consistent and easier to
understand.