Ignore case when checking spam in SpamAddress
Sometimes spammers use different cases for their spam emails. This
allows `SpamAddress.spam?` to be case insensitive so that we don't have
to create multiple records for the same address.

The `email_address` argument can be an `Array` of addresses (we pass the
email `To:` header to this), so this implementation ensures we're always
dealing with an Array and then checks each email against the list.

Fixes https://github.com/mysociety/alaveteli/issues/5624.