Ignore case when checking spam in SpamAddress
Sometimes spammers use different cases for their spam emails. This
allows `SpamAddress.spam?` to be case insensitive so that we don't have
to create multiple records for the same address.

The `email_address` argument can be an `Array` of addresses (we pass the
email `To:` header to this), so this implementation ensures we're always
dealing with an Array and then checks each email against the list.

We could use the Postgres citext [1] extension here, but enabling in a
migration requires the database user to have `SUPERUSER` permissions,
otherwise the migrations fail.

We could opt to require a manual step to enable citext manually, but
that adds extra upgrade complication for reusers.

The easy approach is to just ensure we downcase emails as we save and
compare them.

Fixes https://github.com/mysociety/alaveteli/issues/5624.