Add random info request to render
Render a random InfoRequest from the project to allow us to work on the
templates. We don't yet know how to approach picking from the queue, so
this allows us to make some progress while the feature is in
development.

We also need to set many of the instance variables for the dependent
partials.