Add stub Projects::ClassifiesController
The controller is separate to Project::ClassificationsController because
the latter requires an InfoRequest#url_title param. We could probably
work around it, but it feels like the “queue” controllers are beginning
to develop their own identity which can be generalised later.

Currently just renders a random, classifiable InfoRequest from the
Project.