Set-Cookie must be a String
There's a long history of StripEmptySessions.

It was first introduced in FixMyTransport [1] in 2011 and later ported
to Alaveteli [2].

As far as I can tell, it is invalid to assign an `Array` to the
`Set-Cookie` headers. The current rack SPEC [3] states:

> The values of the header must be Strings, consisting of lines (for
> multiple header values, e.g. multiple Set-Cookie values) separated by
> “\n”

This is also true in the version of rack Alaveteli depends on at the
time of this commit [4].

Since 2.1.0, `Rack::MockRequest` explicitly parses the `Set-Cookie`
header by splitting on `\n` [5] causing spec failures when we bump the
gem.

This commit adds the `lint: true` option when using `Rack::MockRequest`
so that non-compliance with the Rack SPEC [3] raises errors.

This in particular catches the issue this commit is addressing:

    1) StripEmptySessions if there is no meaningful data in the session should strip the session cookie setting header (but no other cookie setting header) if there is more than one
       Failure/Error: response = Rack::MockRequest.new(app).get('/', 'HTTP_ACCEPT' => 'text/html', lint: true)

       Rack::Lint::LintError:
         a header value must be a String, but the value of 'Set-Cookie' is a Array
       # /home/vagrant/bundle/ruby/2.5.0/gems/rack-2.0.8/lib/rack/lint.rb:20:in `assert'
       # /home/vagrant/bundle/ruby/2.5.0/gems/rack-2.0.8/lib/rack/lint.rb:645:in `block in check_headers'
       # /home/vagrant/bundle/ruby/2.5.0/gems/rack-2.0.8/lib/rack/lint.rb:629:in `each'
       # /home/vagrant/bundle/ruby/2.5.0/gems/rack-2.0.8/lib/rack/lint.rb:629:in `check_headers'
       # /home/vagrant/bundle/ruby/2.5.0/gems/rack-2.0.8/lib/rack/lint.rb:53:in `_call'
       # /home/vagrant/bundle/ruby/2.5.0/gems/rack-2.0.8/lib/rack/lint.rb:37:in `call'
       # /home/vagrant/bundle/ruby/2.5.0/gems/rack-2.0.8/lib/rack/mock.rb:74:in `request'
       # /home/vagrant/bundle/ruby/2.5.0/gems/rack-2.0.8/lib/rack/mock.rb:56:in `get'
       # ./spec/lib/strip_empty_sessions_spec.rb:12:in `make_response'
       # ./spec/lib/strip_empty_sessions_spec.rb:67:in `block (3 levels) in <top (required)>'

Therefore, I think changing the specs to not pass an Array is the
correct thing to do here.

This means we can really simplify `StripEmptySessions`.

Instead of type-checking the value of the `Set-Cookie` header we can
always assume it's a String and `gsub!` it as per the original `String`
implementation.

We do now have to handle the fact that `Set-Cookie` may be `nil`, so
this commit adds a spec for that and fixes it by using the safe
navigation operator (`&.`) to only `gsub!` if the header exists.

Also cleans up the conditional by making it single-level.

[1] https://github.com/mysociety/fixmytransport/commit/5b426438
[2] https://github.com/mysociety/alaveteli/commit/30bba84b
[3] https://github.com/rack/rack/blob/v2.2.2/SPEC.rdoc#label-The+Headers
[4] https://github.com/rack/rack/blob/2.0.8/SPEC#L234-L236
[5] https://github.com/rack/rack/blob/2.1.0/lib/rack/mock.rb#L218