Add project classification form
Add a slightly stripped-down `describe_state` form for classifying in
the context of a Project.

Uses the non-pro state transitions so that we get more descriptive form
labels.

Adds a default `id_suffix` to `request/_state_transition` so that we
don't have to explicitly pass a `nil` value if we're only rendering a
single classification form on the page.