Add Project#invite_token
Allows us to find a Project based on a token rather than ID and then add
contributors to the project.

Ideally we'd create an "invitable" concern with a polymorphic model so
that many models can have invite codes, but this is the simplest
solution for now.

The `invite_token` should be manually generated through the rails
console for now with `SecureRandom.uuid`.