Projects::ClassificationsController requires url_title
Explicitly require `url_title` param in
`Projects::ClassificationsController`.

Prior to this commit, submitting a classification message via
`Projects::ClassificationsController` would fail because most other uses
already have a `url_title` in the routes.

This commit uses strong params to require a `:url_title` param so that
the error message is clearer, and adds a hidden field to the form to
ensure the param is submitted if it isn't already submitted via the
routing.