Drop overall project progress bar
Feels like this just adds noise at the moment.

"Overall" progress is essentially "requests with data extracted", so
this would end up showing that. We could try to figure out progress in
terms of number of classifable + extractable requests left to do, but
really its just clearer to push attention down to the task queues.