Remove the project nav from the project homepage
Feels a bit weird that it repeats the title and "contact owner" is
already pretty prominent on that page. Only thing we lose is the link to
"tasks". I think they are fairly obvious as you’re going to just scroll
down the page and see the progress bars.

We do lose that little hint that "hey, you’re on this page and there’s
something for you to do", but maybe we can think of a better way of
doing that instead of the nav.