Extract request/_correspondence event partials
Reduces conceptual complexity. Now we just need to understand that we’re
rendering a named partial per event, rather than having to parse out
which part of the conditional is happening for a given event.

Adding new event types to the stream is much easier now. We just create
a new partial in `app/views/request/events`.

Rescuing from `ActionView::MissingTemplate` allows us to preserve
existing behaviour of not _needing_ a view template for _all_ events. We
don't display anything for a `set_embargo` event, for example.