Don't show correspondence footer in project queues
Using the link to this links via projects won't direct people to the
correct location – it will direct them outside the project.

We also don't really need people to report requests, since they'll all
effectively be "owner" in the context of a project.

Also fixes issue where a portion of the correspondence footer is shown
for embargoed requests, even though we're not rendering it.

Also hides correspondence footer when making a followup. Had to add the
key here so may as well make it `false` as showing the footer doesn't
make sense in this context.