Use Project::Queue
Pull requests to classify from the queue

We don’t need to clear the current request on submission, because that
request will no longer be in the info_requests for the queue. We’ll try
to look it up again, but it won’t be there, so it falls back to sampling
a new random request from the queue. If there are none left, we get
redirected back to the project homepage.