Fix project queue class loading for older Ruby
For Ruby 2.4 and prior...

In the spec failure output, when attempting to initialise
Project::Queue::Classifiable, we were seeing:
  NoMethodError: undefined method `new' for Classifiable:Module

This was due to Thread::Queue existing as a top level constant. Later
Rubies aren't affected due to this change [1].

This can be resolved by using require_dependency when these classes are
referenced.

[1] https://github.com/ruby/ruby/commit/44a2576