Minor spec refactoring
Use extractable_requests_count to avoid needing to mess with the request
state in setup.

This doesn’t actually work correctly now, but the hacks are made clearer
by the code comments and will eventually be pushed back into the
factory.

Also creates a `Project::Submission` for an “extracted” request to be
more aligned with production behaviour.