Create InfoRequest before creating InfoRequestEvent
FactoryBot 5.0.0 sets `use_parent_strategy` to `true` by default [1].

This is resulting in two calls to `create_or_update_request_summary` as
the `InfoRequest` association is saved during the `event.save` call.

I think previously the `InfoRequest` would have been created when the
`InfoRequestEvent` was built, prior to applying the `expect`.

Now we create the `InfoRequest` and pass it to the factory so that
`event.save` only triggers the callback that would happen in normal
operation when a new `InfoRequestEvent` is created.

[1] https://github.com/thoughtbot/factory_bot/blob/master/NEWS.md#500-february-1-2019