Install development dependencies on Travis
We rely quite heavily on dependabot's automated gem updates. However,
it's easy to miss the fact that some development dependencies aren't
mergeable because they aren't installed when running tests, and the
build is green.

This will make builds error when a gem can't be installed due to a Ruby
version constraint. It shouldn't affect performance as we cache gems.