Remove time-stubbing
Just create an `:expiring_embargo` and calculate the new expiry date.
Reduces complexity and makes it easier to remove Delorean, as
`ActiveSupport::Testing::TimeHelpers` don't work in quite the same way.

When trying to use `ActiveSupport::Testing::TimeHelpers`, the stub
wasn't getting applied in the same way so the `Embargo` submitted to the
controller wasn't actually expiring soon. This change was easier than
figuring out exactly what the problem was.

This commit uses the `be_within` matcher as there's a tiny delay between
submitting the controller action and calculating the expected time.