Update expiring embargo factory
Set `#publish_at` to the beginning of the day. This is what we would
expect to be set using the `AlaveteliPro::Embargo.*_months_from_now`
helpers.

It reduces complexity and makes it easier to remove Delorean, as
`ActiveSupport::Testing::TimeHelpers` don't work in quite the same way.
`TimeHelpers` will freeze time by stubbing `Time.now` so time doesn't
progress and also sets usec to be zero [1].

Freezing time at the point an embargo is due to expire causes time
comparison in `AlaveteliPro::Embargo#expiring_soon?` to break as the
`Time.now` will always be less than `#publish_at`.

[1] https://github.com/rails/rails/blob/5-2-stable/activesupport/lib/active_support/testing/time_helpers.rb#L101-L103