Add temp:identify_broken_binary_censor_rules
Prints the URL of each binary attachment that has a censor rule that
cannot be applied due to incompatible character encodings.

Recommended to run with `tee` to monitor progress and save URLs to a
file:

    bundle exec rake temp:identify_broken_binary_censor_rules | \
      tee broken-attachments.txt

Attachment parsing errors will be printed to STDERR, so not included in
the output file, but that's fine as they weren't being parsed anyway.

This task took several hours to run on WhatDoTheyKnow, so recommended it
gets run in a tmux or screen session.

[1] https://github.com/mysociety/alaveteli/issues/3546