Disable Naming/AccessorMethodName for specific methods
These aren't true writer methods – they're configuration.

A writer method would be `AlaveteliLocalization#default_locale=`, but
this isn't writing to an instance variable on `AlaveteliLocalization`.

These probably _should_ be writer methods at some point, but right now
this class is more of a configuration helper rather than an object that
can be initialized.