Fix the locale switcher current locale
This changes ordering before action locale callbacks

We need `set_gettext_locale` to run before `collect_locales` this is
because the URL locale from the params is passed into
`AlaveteliLocalization.set_session_locale` and used to set
`FastGettext.locale`.

This is used then used, via `AlaveteliLocalization.locale` in the
`collect_locales` callback to set the `@locales` varibale used when
render the locale switcher.