Ensure the underscore locale is included in i18n fallbacks
When a site has an underscore locale (e.g. "fr_BE") set as the default,
it is not included in the I18n fallbacks for non-default locales.

    I18n.fallbacks
    # => { :fr_BE => [:fr_BE, :"fr-BE", :fr],
    #      :nl_BE => [:nl_BE, :"fr-BE", :fr] }

    AlaveteliLocalization.with_locale(:fr_BE) { Globalize.fallbacks }
    # => [:fr_BE, :"fr-BE", :fr]
    AlaveteliLocalization.with_locale(:nl_BE) { Globalize.fallbacks }
    # => [:nl_BE, :"fr-BE", :fr]

In the example above, this means that a missing model translation when
using the "nl_BE" locale will **not** correctly fall back to the
translation for the default locale ("fr_BE").

This commit ensures that the default locale is always available in the
fallbacks for non-default locales. We try to add it before the
non-underscore version so that the regional translation is preferred,
with the non-regional fallback available as a last-ditch attempt to find
a useful translation.

Globalize model attributes fallbacks specs contributed by @gbp.

Fixes https://github.com/mysociety/alaveteli/issues/5382
Fixes https://github.com/mysociety/alaveteli/issues/5895