Remove translation markup from admin interface
The admin interface for the site mostly appears in English, but some
strings have been marked up for translation, resulting in a mix of
English and the locale language appearing.

This commit cleans up the translated strings to reduce the amount of
translation that needs to be done and clarifies the interface.

A few cases of translation markup still remain where the string ends up
being seen by a user, most notably in setting default form fields that
get sent to the user by email.

Fixes https://github.com/mysociety/alaveteli/issues/1353.