Allow setting log level through the environment
Sometimes it's useful to be able to use a non-default log level.

This commit allows any log level to be set through `LOG_LEVEL` in the
environment.

If an invalid log level is set the application will fail to boot with a
`NameError`, e.g:

    $ LOG_LEVEL=foobarbaz bin/rails runner "puts Rails.application.config.log_level"
    # snip…
    uninitialized constant ActiveSupport::Logger::FOOBARBAZ (NameError)

If `LOG_LEVEL` is not set, the default for that environment will be
used.