Extract controller code to RawEmail#domain_from_email
For some reason we were accessing `from_email` via the
`IncomingMessage`, but `IncomingMessage#from_email` just delegates to
its `RawEmail`.

Instead, add `RawEmail#domain_from_email` and just call it directly.
This means it gets tested and reduces complexity in the controller.

Would like to avoid the dependency on `PublicBody`, but that's for
another day.