Truncate search queries that are too long
The maximum length of a search query is 252 bytes. People often paste
long strings into the typeahead search which results in an exception
being raised.

Instead of raising an exception, just truncate the query.

We could tell the user the string is too long, but I don't think it's
worth it in this case. The typeahead search is pretty opaque to users –
it just happens as they enter content – and with a string that long I
don't think we'll lose much by truncating it.

Fixes https://github.com/mysociety/alaveteli/issues/5815.