Prevent overflowing of long censor rules
Long text or replacement attributes would overflow the table cell and
overlap other parts of the UI, making it difficult to understand the
rules applied without editing each rule.

This commit forces long strings to wrap, and uses monospace text for
clearer presentation of the data that gets used in the replacement
process.