Improve resilience to unparsable outlook attachments
Sometimes we receive messages with outlook attachments that can't be
parsed due to an issue in mapi [1].

There's a potential fix [2] but it conflicts [3] with an existing patch
we apply [4].

This at least allows users to download the raw attachment, rather than
us preventing the entire request from loading because we raise an
exception.

It's not easy to include an attachment in the specs to replicate a real
error case due to the complexity of removing PII, so I've just stubbed
the call to `.open` as we don't care about the specifics in this spec.

`script/handle-mail-replies` needs an explicit require as we minimise
the load path for that script.

Part of https://github.com/mysociety/alaveteli/issues/5783.

[1] https://github.com/aquasync/ruby-msg/issues/15
[2] https://github.com/aquasync/ruby-msg/pull/16
[3] https://github.com/mysociety/alaveteli/issues/5783#issuecomment-655740838
[4] https://github.com/mysociety/ruby-msg/pull/3