Logically group individual requires
We try to minimise what we load in `script/handle-mail-replies` as it's
run pretty frequently, so we want to avoid memory capacity problems.

This commit does some minor cleanup to group the addition of a directory
to the load path, the requires for the files under that load path, and
any code called directly from the require.

This is to make a future change easier. I've also cleaned up quote
style.