Fix escaped <code> tags in admin interface
Translation markup was removed from the admin interface in 52fd17c, but
this string was relying on the translation being marked as HTML safe.

We could use `raw` to render the HTML contained within the string, but
it's cleaner to get rid of the string altogether.

Fixes https://github.com/mysociety/alaveteli/issues/5924.