Detect legislation references in responses
Adds `IncomingMessage#refusals` method to return any refusals referenced
in the incoming message response. This can be used in the views to
render suggestions to requesters if they've received a refusal.

Requires `Legislation` instance to be initialised with an array refusals
EG `Legislation.new(key: 'foi', refusals: 's 12')`, this will be done in
themes.