Tweak change request button colours
**admin_public_body_change_requests/edit**

"Close and respond" is more frequently used than "Close", so make it the
primary button.

"Close" is not a destructive action, so just make it a standard button.

**admin_general/index**

Mirror the above changes to the buttons in this view.

We don't want to "primary" buttons, so this makes "Add authority" a
"success" button (green) as that's a clear indication of accepting the
change, with "Close and respond" being the primary fallback option.

Fixes https://github.com/mysociety/alaveteli/issues/5933.