Force mail data to have CRLF line endings
Since upgrading to Mail 2.7.1 (c581d83) we've noticed a regression in
handling some PDFs. Instead of being extracted to an FoiAttachment, the
base64 encoded string gets rendered into the main part of the
correspondence. We've isolated the cause to mikel/mail@21222e1.

Since the introduction of Action Mailbox Rails parses raw emails through
Mail through `Mail::Utilities.binary_unsafe_to_crlf` [1]. This appears
to have been introduced in 627bbd34 with no specific explanation for the
use of the helper.

The issue appears with the presence of non-ASCII characters somewhere in
the raw email text of multipart messages.

Mail uses a Regexp to split parts, which only spots \r\n, and thus will
fail on any LF-ended non-ASCII email leaving the whole email as
one non-MIME part. Mail does convert the raw email to CRLF, but only
when the String has a valid encoding, or if the String is encoded as
`Encoding::BINARY`, then only if the String contains only ASCII
characters.

Alaveteli forces all raw emails to `Encoding::BINARY` as of 0f9d143db9,
so `Mail::Utilities.safe_for_line_ending_conversion?` returns `false`.
If the email is ASCII only, line endings are being converted back to
CRLF and it works fine.

Unilaterally forcing a conversion to CRLF, given the mail code parsing
mandates CRLF and isn't always doing the conversion, seems like the
right thing to do.

As well as fixing the above issue, some cleanup is included in this
commit:

* Rely on `mail_from_raw_email` converting to CRLF rather than in
  `email_helpers` to ensure consistent behaviour.
* Reset all email fixture files back to Unix line endings for
  consistency.

Fixes https://github.com/mysociety/alaveteli/issues/5905.

* Initial problem description, specs and cleanup contributed by
  @garethrees
* Bugfix contributed by @gbp
* Conversion issues in Mail identified by @dracos

[1] https://github.com/rails/rails/blob/a5b2fff/actionmailbox/lib/action_mailbox/mail_ext/from_source.rb#L5