Refactor deployments into Jekyll data file
Details about international deployments are now stored in a file at
`_data/deployments.yml` which includes both English and Spanish
translations of country names and deployment descriptions. This should
make deployments easier to update (no more editing HTML!) and make it
less likely for the English and Spanish pages to fall out of sync.

I also took this opportunity to replace the creaky grid system we had
in place for the deployments page, which relied on `display: inline`
and HTML comments to avoid whitespace. We now use a Bootstrap-inspired
grid system, with four breakpoints, floats, and an upgrade to flex
columns for browsers that support it.