Always configure refusal advice paths
Without this change and when not using the theme refusal advise couldn't
be loaded resulting in an exception on the /help/unhappy page where we
try to load the refusal wizard.

This changes the initializer to always setup the path and when requiring
a theme it adds the theme's config/refusal_advice path.

In the test environment this allows us to dynamically add and remove
fixture paths for different sets of specs.

Fixes #6006