Add signal handling to Xapian update wrapper script
This traps a SIGTERM sent to the bash script that wraps the call to rake
to update the Xapian index. We need to ensure that any children are
killed so we use a SIGKILL to do this before exiting.

This means that should the update process hang, a SIGTERM can be sent to
this wrapper and the children will be handled. Given the wrapper may be
called from cron or similar, this provides a simple interface to cleaning
up.