Don't cache backpage attachments
We need to serve the attachment from Rails in order to set the
X-Robots-Tag added in 8e7ac48bea.

Spec is currently skipped because the attachment seems to be already
cached prior to the spec running, so something to investigate and fix.

Need this commit to resolve a case though.