Render snippet library
Work towards #5911.

If you’re viewing the site through a theme that’s overridden the
followup template (eg: whatdotheyknow-theme) you’ll need to make sure
it has the `<%= render :partial => 'snippets' %>` line in it, to render
the snippet library.

Still quite a few TODOs in the code – mostly around integrating the
real data, and tying it into the refusal checklist flow.

We may want to implement a fixed max-height and `overflow: auto` on
the snippet list, once there are lots of snippets to display. My
instinct is that the list *shouldn’t* scroll when we’re displaying the
snippets identified as outputs of the refusal checklist, since we
*know* they will all be relevant to the user. But when the user is
just browsing the other filtered lists of snippets, scrolling is fine.

Also includes JavaScript filtering for snippets.