Fix IncomingMessage#refusals returning duplicates
I had thought the issue was that we didn't have a `Legislation#==`
definition (a6dfc49dc), but that was masking the true problem.

Because `#uniq` method works by making each element a Hash key (which
are unique) [1], we were ending up with different Hash keys even though
semantically our "parents" are the same.

I'm not sure this is the most satisfactory thing to do, but it works
well enough for now. Basically we say "at the end of the day I'm going
to convert these legislation references to a string, so `uniq` on that."

Also gives doubles unique names for two reasons.

1. The change caused the "returns references which are refusals" spec to
   fail since – I assume – calling `#to_s` on the double was returning
   the non-unique names ("refusals")
2. It makes it easier to understand which items are missing/included in
   an array comparison.

Fixes https://github.com/mysociety/alaveteli/issues/6020.

[1] https://www.rubyguides.com/2019/08/ruby-uniq-method/