Add InfoRequest#reason_to_be_unhappy?
Nicer language for checking if a request is in a state where we think
they have reason to be unhappy with the response. If a request has been
classified in one of the unhappy states then we send people to a
specific help page to help them get a better answer.