Fix broken password updates with OTP specs
Since upgrading to active_model_otp 2.1.0 specs for updating password
have been broken.

This is due a bug being fixed upstream [1]. This has resulted in the
user record being saved an additional time [2].

We call `authenticate_otp` within a validate callback it means it's also
called an additional time.

This second call then fails due to the `entered_otp_code` is set to nil
in the initial validation callback. This adds an error to the user
prevents the controller from following the expected successful
redirection.

This change prevents additional save from being triggered upstream and
instead we manually increment the `otp_counter` which is saved after
validation is complete.

[1] https://github.com/heapsource/active_model_otp/issues/58
[2] https://github.com/heapsource/active_model_otp/blob/v2.1.0/lib/active_model/one_time_password.rb#L52