Highlight prominence in admin request lists
Adds an `Admin::ProminenceHelper` to group prominence-related helpers.

Adds `prominence_icon` to render an icon with appropriate styling based
on the prominence. Accepts a String or object that responds to
`#prominence`.

There are probably other list partials that this can be added to, but
this is a good first addition. We can also use this for the `eye` links
that we use to link to the public version of the request, but we'll need
to refactor for that.

Fixes https://github.com/mysociety/alaveteli/issues/4915.