Dynamically render allow_new_responses_from partial
Reduce complex logic in `request/_request_subtitle` by using the
`InfoRequest` attribute to render the correct partial.

To keep things simple, I've also created an empty partial for `anybody`.
We could add a helper [1] to render a template only if it exists, but
we don't expect the set of values to change much. An
`ActionView::MissingTemplate` error is probably the correct result here
if we did introduce a new possible value.

[1] https://coderwall.com/p/ftbmsa/render-template-if-exists-in-rails