Add InfoRequest::TitleValidation to InfoRequestBatch
An `InfoRequestBatch` creates one or more `InfoRequest` records, so the
title should be validated against the same conditions.

We currently ensure this at the controller level by validating the first
`InfoRequest` belonging to the `InfoRequestBatch`, but being confident
that an `InfoRequestBatch` will create valid records could reduce
complexity in controllers – always a good thing.