Undef Kernel.open before creating open scope
All Ruby objects inherit from Object, which includes the Kernel module,
which in turn, has an open method. Redefining `.open` was creating lots
of WARN level messages in the logs:

    W, [2019-10-10T14:17:40.179390 #30917]  WARN -- : Creating scope
    :open. Overwriting existing method PublicBodyChangeRequest.open

Since we're not using `Kernel.open`, we can explicitly undefine it prior
to defining our own `.open` in the form of a `scope`.

Fixes https://github.com/mysociety/alaveteli/issues/5405.