Handle unconfigured legislation questions
`RefusalAdvice#questions` expects a `Legislation` to be configured. It
allows empty questions, but currently we have not even configured EIR.
This leads to failures when we try to render questions for an EIR
request.

This commit removes the need for configuring a `Legislation` with empty
questions by defining `RefusalAdvice::Store#dig` and using that to
return `nil` when we attempt to find questions for an unconfigured
`Legislation`. The existing code then handles converting the question
data to an `Array` and mapping over the empty collection.

Fixes https://github.com/mysociety/alaveteli/issues/6246.