Handle unconfigured legislation actions
`RefusalAdvice#actions` expects a `Legislation` to be configured. It
allows empty actions, but currently we have not even configured EIR.
This leads to failures when we try to render actions for an EIR request.

This commit removes the need for configuring a `Legislation` with empty
actions by using `RefusalAdvice::Store#dig` and using that to return
`nil` when we attempt to find actions for an unconfigured
`Legislation`. The existing code then handles converting the action
data to an `Array` and mapping over the empty collection.