Truncate list of alternative users shown on profiles
The list is becoming unusable as no one can tell the difference between
each user without clicking on them.

For users with a very common name (say for example, John Smith) we
render a listing of nearly all users with the same name before showing
the user's own profile, which results in a very large "wall of text".

This commit removes the listing of accounts and instead links to a user
search of the name.

Fixes https://github.com/mysociety/alaveteli/issues/1732