Remove Admin URL from requires admin notifications
I frequently wonder whether to remove this when I reply to user support
requests. Most of the time I don't bother, since one needs authorisation
to view the page, but it still doesn't _look_ good to reveal to
outsiders.

We also tend to redact these URLs from SAR responses, which creates
significant extra work.

This commit removes the URL. It adds an extra click for admins who need
to act on the request, but personally speaking I almost always use the
public URL first to review the content before visiting the admin page to
act.