Refactor max length InfoRequest::TitleValidation
Define `InfoRequest::TitleValidation` in terms of
`InfoRequest::DraftTitleValidation`.

This reduces duplication in validations. What we're implying here is
that an `InfoRequest::TitleValidation` is everything from the looser
`InfoRequest::DraftTitleValidation` definitions _plus_ the extra
constraints defined here.

I've retained the duplication in the specs, since it's not true to say
an `InfoRequest` behaves like an `InfoRequest::DraftTitleValidation`.