Use number_field for numeric Dataset::Key
This is more appropriate for the expected value and makes the intent
clearer. We also get HTML validations for free.

At present neither the HTML validations nor the `Dataset::Key::FORMATS`
allow decimal values. We could update the format `Regexp` and add a
`step` attribute to the form field if we wanted to accept decimal
values.

Part of https://github.com/mysociety/transparency-fnf-whatdotheyknow-projects/issues/30