Make navigation to project homepage clearer
The arrow emphasises that you're going "back" to the homepage. Spotted
after watching a new user struggle to find their way back to the
homepage.

The initial version I tried used the `&larr;` outside of the link, but
it felt a bit disconnected so I thought it was worth the minor extra
complexity of interpolation.