Remove commented Bing Crawl-delay
We're receiving a "Crawl politeness" notification on WDTK because Bing
is interpreting this commented line as an uncommented directive.

The unsubscribe link on the notification emails just takes you to the
homepage and does nothing.

Just going to remove since we're not using the directive.

Fixes https://github.com/mysociety/sysadmin/issues/1454

[1] https://www.bing.com/webmasters/help/crawl-error-alerts-e29a3f3e