Set InfoRequest#law_used within before validation
Use a validation callback as this value depends on the `public_body`
association being set.

This could be `nil` when initialising,resulting in the law used being
set to the default legislation. To prevent this instead of calling
`InfoRequest#legislation` we always delegate to `public_body` if set.

Fixes https://github.com/mysociety/alaveteli/issues/6322