Fix mail forwarding on to both addresses
When a mail is sent To the `CONTACT_EMAIL` and copies in the
`PRO_CONTACT_EMAIL`, the `CONTACT_EMAIL` gets dropped because the
`PRO_CONTACT_EMAIL` is in the list of recipients.

This commit fixes that by forwarding on to both forwarding addresses
where there's been an explicit inclusion of both addresses in the mail.

This is implemented by updating
`MailHandler::ReplyHandler.get_forward_to_address` to return a
pre-composed `String` ready to pass to the `sendmail` command, which
accepts a comma-separated list of recipients.

Fixes https://github.com/mysociety/alaveteli/issues/5649.

[1] https://stackoverflow.com/a/13390926