Create a config/xapian.yml for Vagrant dev install
As described in 47bfdea, sometimes running specs in Vagrant results in
unrecoverable corruption of the Xapian DB.

We're not sure exactly what the problem is, but our best guess is
something to do with slow file sync of the VirtualBox share after the
index is replaced.

Moving the Xapian database outside of the VirtualBox share _seems_ to
prevent the problem, so when provisioning a Vagrant development
environment write a custom `config/xapian.yml` to write the test
databases to `/tmp/xapiandbs`. It doesn't matter that these are
inaccessible from the host machine or get trashed on restart. We only
create this file if it doesn't already exist

This change also allows us to remove the hack introduced in 47bfdea.