Track supported OS ruby versions in CI
The normal practice for installing Alaveteli is to use the Ruby provided
by the operating system distribution. We support a variety of operating
systems for re-installers, so we build against multiple Ruby versions
to ensure that Alaveteli is going to run correctly when installed on any
of the given systems.

Up until 0ab1c1a we explicitly targeted the Ruby versions provided by
the supported operating systems at the time. This commit restores this
synchronisation.

Using the latest release of any given Ruby is an edge case for more
experienced users who manage the Ruby version themselves (like us at
mySociety). We tend to use only the very latest Ruby, so I've included
that here.

For faster CI feedback, it would probably be worth extracting a separate
GitHub Workflow for this "Operating System Support" matrix, and use only
the latest supported Ruby for the first-pass CI build.