Fix extreme deprecation warnings on Ruby 2.7
Until Ruby 2.7.2 [1] the default was for Ruby to omit warnings about
future deprecations. This caused quite a big issue for Rails at the time
[2] so it was reverted and a different migration plan was put in place.

Ubuntu Focal ships with 2.7.0, so this feature results in an extreme
amount of STDERR output and makes some specs fail where we're expecting
no output.

The only way to fix this is to set the particular Ruby options, so this
commit adds those by writing them to the Alaveteli user's bashrc if the
current ruby is one of the afflicted.

The `export` is also wrapped in a conditional so that the line only
affects 2.7.0 and 2.7.1, in case people forget to remove it.

[1] https://bugs.ruby-lang.org/issues/17000
[2] https://discuss.rubyonrails.org/t/new-2-7-3-0-keyword-argument-pain-point/74980