Migrate test database for development installs
We create the test database and migrated the development / production
database through `script/rails-post-deploy`, but in a development
install we're left without a ready-to-go test database. We don't want to
run `script/rails-post-deploy` with `RAILS_ENV=test` since that runs all
sorts of other things, so just migrate the test database as a separate
task.