Only symlink repository when different to destination
After `vagrant up`  the install script creates an
`alaveteli` -> `/home/vagrant/alaveteli` symlink in the repository which
isn't useful either inside the guest or on the host. It also appears as
a new file in git, which one has to remember to remove.

This commit fixes the issue by only creating the symlink if the
`REPOSITORY` and `LINK_DESTINATION` differ.

Fixes https://github.com/mysociety/alaveteli/issues/1384